Struct aws_sdk_dax::client::fluent_builders::CreateCluster

pub struct CreateCluster<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

Implementations

impl<C, M, R> CreateCluster<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub async fn send(
self
) -> Result<CreateClusterOutput, SdkError<CreateClusterError>> where
R::Policy: SmithyRetryPolicy<CreateClusterInputOperationOutputAlias, CreateClusterOutput, CreateClusterError, CreateClusterInputOperationRetryAlias>,

pub fn cluster_name(self, inp: impl Into<String>) -> Self

The cluster identifier. This parameter is stored as a lowercase string.

Constraints:

A name must contain from 1 to 20 alphanumeric characters or hyphens.

The first character must be a letter.

A name cannot end with a hyphen or contain two consecutive hyphens.

pub fn set_cluster_name(self, input: Option<String>) -> Self

pub fn node_type(self, inp: impl Into<String>) -> Self

The compute and memory capacity of the nodes in the cluster.

pub fn set_node_type(self, input: Option<String>) -> Self

pub fn description(self, inp: impl Into<String>) -> Self

A description of the cluster.

pub fn set_description(self, input: Option<String>) -> Self

pub fn replication_factor(self, inp: i32) -> Self

The number of nodes in the DAX cluster. A replication factor of 1 will create a single-node cluster, without any read replicas. For additional fault tolerance, you can create a multiple node cluster with one or more read replicas. To do this, set ReplicationFactor to a number between 3 (one primary and two read replicas) and 10 (one primary and nine read replicas). If the AvailabilityZones parameter is provided, its length must equal the ReplicationFactor.

AWS recommends that you have at least two read replicas per cluster.

pub fn set_replication_factor(self, input: Option<i32>) -> Self

pub fn availability_zones(self, inp: impl Into<String>) -> Self

Appends an item to AvailabilityZones.

To override the contents of this collection use set_availability_zones.

The Availability Zones (AZs) in which the cluster nodes will reside after the cluster has been created or updated. If provided, the length of this list must equal the ReplicationFactor parameter. If you omit this parameter, DAX will spread the nodes across Availability Zones for the highest availability.

pub fn set_availability_zones(self, input: Option<Vec<String>>) -> Self

pub fn subnet_group_name(self, inp: impl Into<String>) -> Self

The name of the subnet group to be used for the replication group.

DAX clusters can only run in an Amazon VPC environment. All of the subnets that you specify in a subnet group must exist in the same VPC.

pub fn set_subnet_group_name(self, input: Option<String>) -> Self

pub fn security_group_ids(self, inp: impl Into<String>) -> Self

Appends an item to SecurityGroupIds.

To override the contents of this collection use set_security_group_ids.

A list of security group IDs to be assigned to each node in the DAX cluster. (Each of the security group ID is system-generated.)

If this parameter is not specified, DAX assigns the default VPC security group to each node.

pub fn set_security_group_ids(self, input: Option<Vec<String>>) -> Self

pub fn preferred_maintenance_window(self, inp: impl Into<String>) -> Self

Specifies the weekly time range during which maintenance on the DAX cluster is performed. It is specified as a range in the format ddd:hh24:mi-ddd:hh24:mi (24H Clock UTC). The minimum maintenance window is a 60 minute period. Valid values for ddd are:

sun

mon

tue

wed

thu

fri

sat

Example: sun:05:00-sun:09:00

If you don't specify a preferred maintenance window when you create or modify a cache cluster, DAX assigns a 60-minute maintenance window on a randomly selected day of the week.

pub fn set_preferred_maintenance_window(self, input: Option<String>) -> Self

pub fn notification_topic_arn(self, inp: impl Into<String>) -> Self

The Amazon Resource Name (ARN) of the Amazon SNS topic to which notifications will be sent.

The Amazon SNS topic owner must be same as the DAX cluster owner.

pub fn set_notification_topic_arn(self, input: Option<String>) -> Self

pub fn iam_role_arn(self, inp: impl Into<String>) -> Self

A valid Amazon Resource Name (ARN) that identifies an IAM role. At runtime, DAX will assume this role and use the role's permissions to access DynamoDB on your behalf.

pub fn set_iam_role_arn(self, input: Option<String>) -> Self

pub fn parameter_group_name(self, inp: impl Into<String>) -> Self

The parameter group to be associated with the DAX cluster.

pub fn set_parameter_group_name(self, input: Option<String>) -> Self

pub fn tags(self, inp: impl Into<Tag>) -> Self

Appends an item to Tags.

To override the contents of this collection use set_tags.

A set of tags to associate with the DAX cluster.

pub fn set_tags(self, input: Option<Vec<Tag>>) -> Self

pub fn sse_specification(self, inp: SseSpecification) -> Self

Represents the settings used to enable server-side encryption on the cluster.

pub fn set_sse_specification(self, input: Option<SseSpecification>) -> Self

pub fn cluster_endpoint_encryption_type(
self,
inp: ClusterEndpointEncryptionType
) -> Self

The type of encryption the cluster's endpoint should support. Values are:

NONE for no encryption

TLS for Transport Layer Security

pub fn set_cluster_endpoint_encryption_type(
self,
input: Option<ClusterEndpointEncryptionType>
) -> Self