Module aws_sdk_dax::error

Modules

cluster_already_exists_fault

See ClusterAlreadyExistsFault

cluster_not_found_fault

See ClusterNotFoundFault

cluster_quota_for_customer_exceeded_fault

See ClusterQuotaForCustomerExceededFault

insufficient_cluster_capacity_fault

See InsufficientClusterCapacityFault

invalid_arn_fault

See InvalidArnFault

invalid_cluster_state_fault

See InvalidClusterStateFault

invalid_parameter_combination_exception

See InvalidParameterCombinationException

invalid_parameter_group_state_fault

See InvalidParameterGroupStateFault

invalid_parameter_value_exception

See InvalidParameterValueException

invalid_subnet

See InvalidSubnet

invalid_vpc_network_state_fault

See InvalidVpcNetworkStateFault

node_not_found_fault

See NodeNotFoundFault

node_quota_for_cluster_exceeded_fault

See NodeQuotaForClusterExceededFault

node_quota_for_customer_exceeded_fault

See NodeQuotaForCustomerExceededFault

parameter_group_already_exists_fault

See ParameterGroupAlreadyExistsFault

parameter_group_not_found_fault

See ParameterGroupNotFoundFault

parameter_group_quota_exceeded_fault

See ParameterGroupQuotaExceededFault

service_linked_role_not_found_fault

See ServiceLinkedRoleNotFoundFault

service_quota_exceeded_exception

See ServiceQuotaExceededException

subnet_group_already_exists_fault

See SubnetGroupAlreadyExistsFault

subnet_group_in_use_fault

See SubnetGroupInUseFault

subnet_group_not_found_fault

See SubnetGroupNotFoundFault

subnet_group_quota_exceeded_fault

See SubnetGroupQuotaExceededFault

subnet_in_use

See SubnetInUse

subnet_quota_exceeded_fault

See SubnetQuotaExceededFault

tag_not_found_fault

See TagNotFoundFault

tag_quota_per_resource_exceeded

See TagQuotaPerResourceExceeded

Structs

ClusterAlreadyExistsFault

You already have a DAX cluster with the given identifier.

ClusterNotFoundFault

The requested cluster ID does not refer to an existing DAX cluster.

ClusterQuotaForCustomerExceededFault

You have attempted to exceed the maximum number of DAX clusters for your AWS

CreateClusterError

CreateParameterGroupError

CreateSubnetGroupError

DecreaseReplicationFactorError

DeleteClusterError

DeleteParameterGroupError

DeleteSubnetGroupError

DescribeClustersError

DescribeDefaultParametersError

DescribeEventsError

DescribeParameterGroupsError

DescribeParametersError

DescribeSubnetGroupsError

IncreaseReplicationFactorError

InsufficientClusterCapacityFault

There are not enough system resources to create the cluster you requested (or to

InvalidArnFault

The Amazon Resource Name (ARN) supplied in the request is not valid.

InvalidClusterStateFault

The requested DAX cluster is not in the available

InvalidParameterCombinationException

Two or more incompatible parameters were specified.

InvalidParameterGroupStateFault

One or more parameters in a parameter group are in an invalid state.

InvalidParameterValueException

The value for a parameter is invalid.

InvalidSubnet

An invalid subnet identifier was specified.

InvalidVpcNetworkStateFault

The VPC network is in an invalid state.

ListTagsError

NodeNotFoundFault

None of the nodes in the cluster have the given node ID.

NodeQuotaForClusterExceededFault

You have attempted to exceed the maximum number of nodes for a DAX

NodeQuotaForCustomerExceededFault

You have attempted to exceed the maximum number of nodes for your AWS

ParameterGroupAlreadyExistsFault

The specified parameter group already exists.

ParameterGroupNotFoundFault

The specified parameter group does not exist.

ParameterGroupQuotaExceededFault

You have attempted to exceed the maximum number of parameter groups.

RebootNodeError

ServiceLinkedRoleNotFoundFault

The specified service linked role (SLR) was not found.

ServiceQuotaExceededException

You have reached the maximum number of x509 certificates that can be created for

SubnetGroupAlreadyExistsFault

The specified subnet group already exists.

SubnetGroupInUseFault

The specified subnet group is currently in use.

SubnetGroupNotFoundFault

The requested subnet group name does not refer to an existing subnet

SubnetGroupQuotaExceededFault

The request cannot be processed because it would exceed the allowed number of

SubnetInUse

The requested subnet is being used by another subnet group.

SubnetQuotaExceededFault

The request cannot be processed because it would exceed the allowed number of

TagNotFoundFault

The tag does not exist.

TagQuotaPerResourceExceeded

You have exceeded the maximum number of tags for this DAX cluster.

TagResourceError

UntagResourceError

UpdateClusterError

UpdateParameterGroupError

UpdateSubnetGroupError

Enums

CreateClusterErrorKind

CreateParameterGroupErrorKind

CreateSubnetGroupErrorKind

DecreaseReplicationFactorErrorKind

DeleteClusterErrorKind

DeleteParameterGroupErrorKind

DeleteSubnetGroupErrorKind

DescribeClustersErrorKind

DescribeDefaultParametersErrorKind

DescribeEventsErrorKind

DescribeParameterGroupsErrorKind

DescribeParametersErrorKind

DescribeSubnetGroupsErrorKind

IncreaseReplicationFactorErrorKind

ListTagsErrorKind

RebootNodeErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateClusterErrorKind

UpdateParameterGroupErrorKind

UpdateSubnetGroupErrorKind