Module aws_sdk_dax::model

Modules

cluster

See Cluster

endpoint

See Endpoint

event

See Event

node

See Node

node_type_specific_value

See NodeTypeSpecificValue

notification_configuration

See NotificationConfiguration

parameter

See Parameter

parameter_group

See ParameterGroup

parameter_group_status

See ParameterGroupStatus

parameter_name_value

See ParameterNameValue

security_group_membership

See SecurityGroupMembership

sse_description

See SseDescription

sse_specification

See SseSpecification

subnet

See Subnet

subnet_group

See SubnetGroup

tag

See Tag

Structs

Cluster

Contains all of the attributes of a specific DAX cluster.

Endpoint

Represents the information required for client programs to connect to the

Event

Represents a single occurrence of something interesting within the system. Some

Node

Represents an individual node within a DAX cluster.

NodeTypeSpecificValue

Represents a parameter value that is applicable to a particular node

NotificationConfiguration

Describes a notification topic and its status. Notification topics are used for

Parameter

Describes an individual setting that controls some aspect of DAX

ParameterGroup

A named set of parameters that are applied to all of the nodes in a DAX

ParameterGroupStatus

The status of a parameter group.

ParameterNameValue

An individual DAX parameter.

SecurityGroupMembership

An individual VPC security group and its status.

SseDescription

The description of the server-side encryption status on the specified DAX cluster.

SseSpecification

Represents the settings used to enable server-side encryption.

Subnet

Represents the subnet associated with a DAX cluster. This parameter refers to

SubnetGroup

Represents the output of one of the following actions:

Tag

A description of a tag. Every tag is a key-value pair. You can add up to 50 tags to a single

Enums

ChangeType

ClusterEndpointEncryptionType

IsModifiable

ParameterType

SourceType

SseStatus