Module aws_sdk_dax::operation

Structs

CreateCluster

Creates a DAX cluster. All nodes in the cluster run the same DAX caching software.

CreateParameterGroup

Creates a new parameter group. A parameter group is a collection of parameters that

CreateSubnetGroup

Creates a new subnet group.

DecreaseReplicationFactor

Removes one or more nodes from a DAX cluster.

DeleteCluster

Deletes a previously provisioned DAX cluster.

DeleteParameterGroup

Deletes the specified parameter group. You cannot delete a parameter group if it is

DeleteSubnetGroup

Deletes a subnet group.

DescribeClusters

Returns information about all provisioned DAX clusters if no cluster identifier

DescribeDefaultParameters

Returns the default system parameter information for the DAX caching

DescribeEvents

Returns events related to DAX clusters and parameter groups. You can obtain

DescribeParameterGroups

Returns a list of parameter group descriptions. If a parameter group name is

DescribeParameters

Returns the detailed parameter list for a particular parameter group.

DescribeSubnetGroups

Returns a list of subnet group descriptions. If a subnet group name is specified,

IncreaseReplicationFactor

Adds one or more nodes to a DAX cluster.

ListTags

List all of the tags for a DAX cluster. You can call ListTags up to

RebootNode

Reboots a single node of a DAX cluster. The reboot action takes place

TagResource

Associates a set of tags with a DAX resource. You can call TagResource up to 5 times per second, per

UntagResource

Removes the association of tags from a DAX resource. You can call

UpdateCluster

Modifies the settings for a DAX cluster. You can use this action to change one or

UpdateParameterGroup

Modifies the parameters of a parameter group. You can modify up to 20

UpdateSubnetGroup

Modifies an existing subnet group.